Integrate nimhax
This is a bit of a hacky way to add nimhax, but it does work.

If movable.sed is not found on the SD card, it will attempt to use
nimhax to dump movable.sed instead. This way, there is no reason to not
have a movable.sed, unless an exploit does not have the ability to run
ctr-httpwn (such as ninjhax 1.)